Pull out results and search sections into their own files
